refactor: refactor scripting tools & added sp
