+++ DbScripts/Tables/FK_Refs_fmis_module_permissions.txt
@@ -1,0 +1,6 @@
+Column 'module_id' in table 'fmis.module_permissions' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Permissions_Modules'.
+Column 'module_id' in table 'fmis.module_permissions' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Permissions_Modules'.
