+++ DbScripts/Tables/FK_Refs_fmis_modules.txt
@@ -1,0 +1,12 @@
+Column 'grand_parent_id' in table 'fmis.modules' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Modules_GrandParent'.
+Column 'parent_id' in table 'fmis.modules' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Modules_Parent'.
+Column 'grand_parent_id' in table 'fmis.modules' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Modules_GrandParent'.
+Column 'parent_id' in table 'fmis.modules' references 
+'fmis.modules'(id) 
+via foreign key 'FK_Modules_Parent'.
