+++ DbScripts/Tables/FK_Refs_fmis_refresh_tokens.txt
@@ -1,0 +1,6 @@
+Column 'user_id' in table 'fmis.refresh_tokens' references 
+'fmis.users'(id) 
+via foreign key 'FK_RefreshTokens_Users'.
+Column 'user_id' in table 'fmis.refresh_tokens' references 
+'fmis.users'(id) 
+via foreign key 'FK_RefreshTokens_Users'.
