+++ DbScripts/Tables/FK_Refs_fmis_role_module_permissions.txt
@@ -1,0 +1,12 @@
+Column 'role_id' in table 'fmis.role_module_permissions' references 
+'fmis.roles'(id) 
+via foreign key 'FK_RolePermissions_Roles'.
+Column 'permission_id' in table 'fmis.role_module_permissions' references 
+'fmis.module_permissions'(id) 
+via foreign key 'FK_RolePermissions_Permissions'.
+Column 'role_id' in table 'fmis.role_module_permissions' references 
+'fmis.roles'(id) 
+via foreign key 'FK_RolePermissions_Roles'.
+Column 'permission_id' in table 'fmis.role_module_permissions' references 
+'fmis.module_permissions'(id) 
+via foreign key 'FK_RolePermissions_Permissions'.
