+++ DbScripts/Tables/FK_Refs_fmis_user_security_questions.txt
@@ -1,0 +1,12 @@
+Column 'user_id' in table 'fmis.user_security_questions' references 
+'fmis.users'(id) 
+via foreign key 'FK_UserSecurityQuestions_User'.
+Column 'security_question_id' in table 'fmis.user_security_questions' references 
+'fmis.security_questions'(id) 
+via foreign key 'FK_UserSecurityQuestions_SecurityQuestions'.
+Column 'user_id' in table 'fmis.user_security_questions' references 
+'fmis.users'(id) 
+via foreign key 'FK_UserSecurityQuestions_User'.
+Column 'security_question_id' in table 'fmis.user_security_questions' references 
+'fmis.security_questions'(id) 
+via foreign key 'FK_UserSecurityQuestions_SecurityQuestions'.
