+++ DbScripts/Tables/FK_Refs_fmis_user_special_permissions.txt
@@ -1,0 +1,6 @@
+Column 'user_id' in table 'fmis.user_special_permissions' references 
+'fmis.users'(id) 
+via foreign key 'FK_UserSpecialPermissions_User'.
+Column 'user_id' in table 'fmis.user_special_permissions' references 
+'fmis.users'(id) 
+via foreign key 'FK_UserSpecialPermissions_User'.
