+++ DbScripts/Tables/FK_Refs_fmis_users.txt
@@ -1,0 +1,6 @@
+Column 'status_id' in table 'fmis.users' references 
+'mwss.key_categories'(id) 
+via foreign key 'FK_Users_StatusKeyCategory'.
+Column 'status_id' in table 'fmis.users' references 
+'mwss.key_categories'(id) 
+via foreign key 'FK_Users_StatusKeyCategory'.
