+++ DbScripts/Tables/key_categories.sql
@@ -1,0 +1,21 @@
+/****** Object:  Table [mwss].[key_categories]    Script Date: 6/3/2025 3:19:52 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [mwss].[key_categories](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[category_name] [nvarchar](255) NOT NULL,
+	[category_id] [int] NULL,
+	[category_value] [nvarchar](255) NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__key_cate__3213E83F82551AA5] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+ALTER TABLE [mwss].[key_categories] ADD  DEFAULT (getdate()) FOR [created_at]
