+++ DbScripts/Tables/roles.sql
@@ -1,0 +1,20 @@
+/****** Object:  Table [fmis].[roles]    Script Date: 6/3/2025 3:18:22 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [fmis].[roles](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[role_name] [nvarchar](100) NOT NULL,
+	[role_status] [bit] NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__roles__3213E83FF091C0D4] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+ALTER TABLE [fmis].[roles] ADD  DEFAULT (getdate()) FOR [created_at]
