+++ DbScripts/Tables/security_questions.sql
@@ -1,0 +1,20 @@
+/****** Object:  Table [fmis].[security_questions]    Script Date: 6/3/2025 3:18:42 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [fmis].[security_questions](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[question_text] [nvarchar](500) NOT NULL,
+	[question_status] [int] NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__security__3213E83F71F37A23] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+ALTER TABLE [fmis].[security_questions] ADD  DEFAULT (getdate()) FOR [created_at]
