+++ DbScripts/Tables/user_security_questions.sql
@@ -1,0 +1,97 @@
+/****** Object:  Table [fmis].[security_questions]    Script Date: 6/3/2025 3:19:12 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [fmis].[security_questions](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[question_text] [nvarchar](500) NOT NULL,
+	[question_status] [int] NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__security__3213E83F71F37A23] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+/****** Object:  Table [mwss].[key_categories]    Script Date: 6/3/2025 3:19:12 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [mwss].[key_categories](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[category_name] [nvarchar](255) NOT NULL,
+	[category_id] [int] NULL,
+	[category_value] [nvarchar](255) NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__key_cate__3213E83F82551AA5] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+/****** Object:  Table [fmis].[users]    Script Date: 6/3/2025 3:19:12 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [fmis].[users](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[first_name] [nvarchar](100) NOT NULL,
+	[middle_name] [nvarchar](100) NULL,
+	[last_name] [nvarchar](100) NOT NULL,
+	[email] [nvarchar](255) NOT NULL,
+	[password] [nvarchar](255) NOT NULL,
+	[status_id] [int] NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__users__3213E83F79457AED] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+/****** Object:  Table [fmis].[user_security_questions]    Script Date: 6/3/2025 3:19:12 PM ******/
+SET ANSI_NULLS ON
+SET QUOTED_IDENTIFIER ON
+CREATE TABLE [fmis].[user_security_questions](
+	[id] [int] IDENTITY(1,1) NOT NULL,
+	[user_id] [int] NOT NULL,
+	[security_question_id] [int] NOT NULL,
+	[security_answer] [nvarchar](255) NOT NULL,
+	[created_at] [datetime2](7) NULL,
+	[created_by] [int] NULL,
+	[updated_at] [datetime2](7) NULL,
+	[updated_by] [int] NULL,
+	[deleted_at] [datetime2](7) NULL,
+	[deleted_by] [int] NULL,
+ CONSTRAINT [PK__user_sec__3213E83FDAD4F491] PRIMARY KEY CLUSTERED 
+(
+	[id] ASC
+)WITH (PAD_INDEX = OFF, STATISTICS_NORECOMPUTE = OFF, IGNORE_DUP_KEY = OFF, ALLOW_ROW_LOCKS = ON, ALLOW_PAGE_LOCKS = ON, OPTIMIZE_FOR_SEQUENTIAL_KEY = OFF) ON [PRIMARY]
+) ON [PRIMARY]
+
+ALTER TABLE [fmis].[security_questions] ADD  DEFAULT (getdate()) FOR [created_at]
+ALTER TABLE [mwss].[key_categories] ADD  DEFAULT (getdate()) FOR [created_at]
+ALTER TABLE [fmis].[users] ADD  DEFAULT (getdate()) FOR [created_at]
+ALTER TABLE [fmis].[user_security_questions] ADD  DEFAULT (getdate()) FOR [created_at]
+ALTER TABLE [fmis].[users]  WITH CHECK ADD  CONSTRAINT [FK_Users_StatusKeyCategory] FOREIGN KEY([status_id])
+REFERENCES [mwss].[key_categories] ([id])
+ALTER TABLE [fmis].[users] CHECK CONSTRAINT [FK_Users_StatusKeyCategory]
+ALTER TABLE [fmis].[user_security_questions]  WITH CHECK ADD  CONSTRAINT [FK_UserSecurityQuestions_SecurityQuestions] FOREIGN KEY([security_question_id])
+REFERENCES [fmis].[security_questions] ([id])
+ON DELETE CASCADE
+ALTER TABLE [fmis].[user_security_questions] CHECK CONSTRAINT [FK_UserSecurityQuestions_SecurityQuestions]
+ALTER TABLE [fmis].[user_security_questions]  WITH CHECK ADD  CONSTRAINT [FK_UserSecurityQuestions_User] FOREIGN KEY([user_id])
+REFERENCES [fmis].[users] ([id])
+ON DELETE CASCADE
+ALTER TABLE [fmis].[user_security_questions] CHECK CONSTRAINT [FK_UserSecurityQuestions_User]
