+++ appsettings.json
@@ -1,6 +1,6 @@
 {
   "ConnectionStrings": {
-    "DefaultConnection": "Server=AL\\SQLEXPRESS;Database=generic_db;Trusted_Connection=True; TrustServerCertificate=True;"
+    "DefaultConnection": "Server=AL\\SQLEXPRESS;Database=simplevia_common;Trusted_Connection=True; TrustServerCertificate=True;"
   },
   "Logging": {
     "LogLevel": {
